+++ extensions/iceworks-material-import/web/mocks/material/getSourcesByProjectType.json
@@ -3,6 +3,6 @@
     "name": "PC Web",
     "type": "react",
     "source": "https://ice.alicdn.com/assets/materials/react-materials.json",
-    "description": "基于 Fusion 基础组件和 ice 脚手架的官方物料"
+    "description": "基于 Fusion 基础组件和 ICE 脚手架的官方物料"
   }
 ]
