+++ extensions/iceworks-material-import/web/package.json
@@ -21,8 +21,7 @@
     "@iceworks/vscode-webview": "^0.1.1",
     "moment": "^2.24.0",
     "react": "^16.8.0",
-    "react-dom": "^16.8.0",
-    "react-sortable-hoc": "^1.11.0"
+    "react-dom": "^16.8.0"
   },
   "scripts": {
     "start": "icejs start --disable-reload",
