--- packages/block-service/src/utils/getIceVersion.ts
@@ -1,5 +1,0 @@
-export function getIceVersion(packageJSON): string {
-  const dependencies = packageJSON.dependencies || {};
-  const hasIceDesignBase = dependencies['@icedesign/base'];
-  return hasIceDesignBase ? '0.x' : '1.x';
-}
