+++ packages/material-utils/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@iceworks/material-utils",
-  "version": "0.1.4",
+  "version": "0.1.5",
   "description": "General tool library for materials.",
   "files": [
     "lib"
