+++ packages/project-service/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@iceworks/project-service",
-  "version": "0.1.8",
+  "version": "0.1.9",
   "description": "Iceworks project service for VSCode extension.",
   "files": [
     "lib"
