+++ packages/vscode-webview/src/vscode.ts
@@ -43,7 +43,7 @@ export function connectService(webview, context: vscode.ExtensionContext, { serv
           });
           // set the optional param to undefined 
           const newArgs = args.concat(Array(api.length - args.length).fill(undefined));
-          const result = args ? await api(...newArgs, context) : await api(context);
+          const result = await api(...newArgs, context);
           console.log('invoke service result', result);
           webview.postMessage({ eventId, result });
         } catch (err) {
